(cGAN) Conditional Generative Adversarial Network/requirements.txt
numpy==1.17.3
Pillow==6.2.1
